requirements.txt
astor==0.8.1
attrs==21.2.0
autograd==1.3
autograd-gamma==0.5.0
Brotli==1.0.9
certifi==2021.5.30
charset-normalizer==2.0.4
chart-studio==1.1.0
click==8.0.1
click-plugins==1.1.1
cligj==0.7.2
cycler==0.10.0
dash==1.21.0
dash-bootstrap-components==0.12.2
dash-core-components==1.17.1
dash-html-components==1.1.4
dash-table==4.12.0
Fiona==1.8.20
Flask==2.0.1
Flask-Compress==1.10.1
formulaic==0.2.4
future==0.18.2
geopandas==0.9.0
idna==3.2
interface-meta==1.2.3
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.2
lifelines==0.26.0
MarkupSafe==2.0.1
matplotlib==3.4.3
munch==2.5.0
numpy==1.21.1
pandas==1.3.1
Pillow==8.3.1
pkg-resources==0.0.0
plotly==5.1.0
pyparsing==2.4.7
pyproj==3.1.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
retrying==1.3.3
scipy==1.7.1
Shapely==1.7.1
six==1.16.0
tenacity==8.0.1
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
